Preliminary break down of wmb components
